+++ familyhelper-note-impl/src/main/java/com/dwarfeng/familyhelper/note/impl/util/FtpConstants.java
@@ -8,8 +8,8 @@ package com.dwarfeng.familyhelper.note.impl.util;
  */
 public final class FtpConstants {
 
-    public static final String PATH_NOTE_FILE = "familyhelper-note/note-file";
-    public static final String PATH_ATTACHMENT_FILE = "familyhelper-note/attachment-file";
+    public static final String[] PATH_NOTE_FILE = new String[]{"familyhelper-note", "note-file"};
+    public static final String[] PATH_ATTACHMENT_FILE = new String[]{"familyhelper-note", "attachment-file"};
 
     private FtpConstants() {
         throw new IllegalStateException("禁止实例化");